Adding makefile for the fibonacci project
Added the makefile for the fibonacci project using g++. Can be called with build
for compiling in optimized form, tests for running it with the unit tests,
and clean to remove *.o and executables. Runs on windows and linux.
Specific for the fibonacci project as there seem to be some
incompatibilities in using wildchar and pathsubs on windows, will be
extended once the linux environment is up and running.